+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * Display the total estimated price of PO in the actual currencies of the PO Lines, rather than in the converted value. Refs UIOR-1264.
 * Submit only fields registered in the form. Refs UIOR-1481.
 * Display a confirmation modal when activating "Bindery active" field in the PO Line form of post-pending order. Refs UIOR-1479.
+* Enhance order template to use a vendor information. Refs UIOR-1482.
 
 ## [8.0.5](https://github.com/folio-org/ui-orders/tree/v8.0.5) (2025-06-30)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v8.0.4...v8.0.5)